helpers: move reference map to separate utils pkg
