+++ tree_test.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/markuskont/go-sigma-rule-engine/utils"
 	"gopkg.in/yaml.v2"
 )
 
@@ -18,7 +19,7 @@ func TestTreeParse(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		var obj DynamicMap
+		var obj utils.Map
 		// Positive cases
 		for _, c := range c.Pos {
 			if err := json.Unmarshal([]byte(c), &obj); err != nil {
@@ -52,7 +53,7 @@ func benchmarkCase(b *testing.B, rawRule, rawEvent string) {
 	if err != nil {
 		b.Fail()
 	}
-	var event DynamicMap
+	var event utils.Map
 	if err := json.Unmarshal([]byte(parseTestCases[0].Pos[0]), &event); err != nil {
 		b.Fail()
 	}
